refactor window boundaries methods
