+++ x/emissions/keeper/msgserver/msg_server_worker_payload.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	errorsmod "cosmossdk.io/errors"
+	"github.com/allora-network/allora-chain/x/emissions/keeper"
 	actorutils "github.com/allora-network/allora-chain/x/emissions/keeper/actor_utils"
 	"github.com/allora-network/allora-chain/x/emissions/metrics"
 	"github.com/allora-network/allora-chain/x/emissions/types"
@@ -65,7 +66,7 @@ func (ms msgServer) InsertWorkerPayload(ctx context.Context, msg *types.InsertWo
 	}
 
 	// Note: this is exclusive of the end block height
-	withinWindow, err := ms.k.BlockWithinWorkerSubmissionWindowOfNonce(topic, *nonce, blockHeight)
+	withinWindow, err := keeper.BlockWithinWorkerSubmissionWindowOfNonce(topic, *nonce, blockHeight)
 	if err != nil {
 		return nil, err
 	} else if !withinWindow {
